Refactored carousel into many components
