--- assets/src/edit-story/components/carousel/carousel.js
@@ -1,530 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import styled, { css } from 'styled-components';
-import { rgba } from 'polished';
-import { useLayoutEffect, useMemo, useRef, useState, useCallback } from 'react';
-
-/**
- * WordPress dependencies
- */
-import { __, sprintf } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import { useConfig, useStory } from '../../app';
-import {
-  LeftArrow,
-  RightArrow,
-  GridView as GridViewButton,
-  Plain,
-} from '../button';
-import {
-  Reorderable,
-  ReorderableSeparator,
-  ReorderableItem,
-} from '../reorderable';
-import Modal from '../modal';
-import useResizeEffect from '../../utils/useResizeEffect';
-import WithTooltip from '../tooltip';
-import KeyboardShortcutsMenu from '../keyboardShortcutsMenu';
-import GridView from './gridview';
-import PagePreview, { THUMB_FRAME_HEIGHT } from './pagepreview';
-import useCarouselKeys from './useCarouselKeys';
-
-const CAROUSEL_BOTTOM_SCROLL_MARGIN = 8;
-
-const MENU_WIDTH = 167;
-const MENU_WIDTH_VERYWIDE = 175;
-const BUTTON_WIDTH = 32;
-
-const Wrapper = styled.section`
-  position: relative;
-  display: grid;
-  grid:
-    '. . prev-navigation carousel next-navigation . menu' auto /
-    ${({ menuWidth }) => menuWidth}px
-    1fr
-    ${BUTTON_WIDTH}px
-    auto
-    ${BUTTON_WIDTH}px
-    1fr
-    ${({ menuWidth }) => menuWidth}px;
-  background-color: ${({ theme }) => theme.colors.bg.workspace};
-  color: ${({ theme }) => theme.colors.fg.white};
-  width: 100%;
-  height: 100%;
-`;
-
-const Area = styled.div`
-  grid-area: ${({ area }) => area};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-`;
-
-const NavArea = styled(Area)`
-  margin-bottom: ${({ marginBottom }) => marginBottom}px;
-  display: flex;
-  justify-content: center;
-`;
-
-const MenuArea = styled(Area).attrs({ area: 'menu' })``;
-
-const EditorVersion = styled.div`
-  display: inline-block;
-  position: absolute;
-  bottom: 0;
-  z-index: 1;
-  margin-left: 14px;
-  margin-bottom: 10px;
-  pointer-events: none;
-  font-size: ${({ theme }) => theme.fonts.version.size};
-  font-family: ${({ theme }) => theme.fonts.version.family};
-  line-height: ${({ theme }) => theme.fonts.version.lineHeight};
-  letter-spacing: ${({ theme }) => theme.fonts.version.letterSpacing};
-  color: ${({ theme }) => rgba(theme.colors.fg.white, 0.3)};
-`;
-
-const PlainStyled = styled(Plain)`
-  background-color: ${({ theme }) => rgba(theme.colors.fg.white, 0.1)};
-  color: ${({ theme }) => rgba(theme.colors.fg.white, 0.86)};
-  &:hover {
-    background-color: ${({ theme }) => rgba(theme.colors.fg.white, 0.25)};
-  }
-`;
-
-const MenuIconsWrapper = styled.div`
-  ${({ isCompact }) =>
-    isCompact
-      ? css`
-          margin-bottom: 16px;
-        `
-      : css`
-          position: absolute;
-          bottom: 44px;
-        `}
-`;
-
-const OverflowButtons = styled.div`
-  position: relative;
-  & > * {
-    position: absolute;
-    bottom: 10px;
-  }
-`;
-
-const StyledGridViewButton = styled(GridViewButton).attrs({
-  height: '24',
-  width: '24',
-})``;
-
-const PageList = styled(Reorderable).attrs({
-  role: 'listbox',
-  'aria-orientation': 'horizontal',
-})`
-  grid-area: carousel;
-  flex-direction: row;
-  align-items: center;
-  justify-content: ${({ hasHorizontalOverflow }) =>
-    hasHorizontalOverflow ? 'flex-start' : 'center'};
-  overflow-x: hidden;
-  overflow-y: hidden;
-  max-width: ${({ maxWidth }) => maxWidth}px;
-  margin: 0 0 ${CAROUSEL_BOTTOM_SCROLL_MARGIN}px 0;
-  padding: 0 16px;
-
-  ${({ hasHorizontalOverflow }) =>
-    hasHorizontalOverflow &&
-    css`
-      overflow-x: auto;
-      overflow-x: overlay;
-      /*
-       * These overrides are an exception - generally scrollbars should all
-       * look the same. We do this only here because this scrollbar is always visible
-       * if scroll is possible.
-       */
-      scrollbar-color: ${({ theme }) => theme.colors.bg.v10}
-        ${({ theme }) => theme.colors.bg.workspace} !important;
-
-      &::-webkit-scrollbar-track {
-        background: ${({ theme }) => theme.colors.bg.workspace} !important;
-      }
-
-      &::-webkit-scrollbar-thumb {
-        border: 2px solid ${({ theme }) => theme.colors.bg.workspace} !important;
-        border-top-width: 3px !important;
-      }
-    `};
-`;
-
-const LINE_WIDTH = 4;
-
-const Line = styled.div`
-  background: ${({ theme }) => theme.colors.accent.primary};
-  height: ${({ height }) => height - THUMB_FRAME_HEIGHT}px;
-  width: ${LINE_WIDTH}px;
-  margin: 0px;
-`;
-
-const ItemContainer = styled.div`
-  display: flex;
-  position: relative;
-`;
-
-const PageSeparator = styled(ReorderableSeparator)`
-  position: absolute;
-  bottom: 0;
-  left: ${({ width }) => width / 2}px;
-  width: ${({ width, margin }) => width + margin}px;
-  height: ${({ height }) => height - THUMB_FRAME_HEIGHT}px;
-  display: flex;
-  justify-content: center;
-
-  ${ItemContainer}:first-of-type &:first-of-type,
-  ${ItemContainer}:last-of-type &:last-of-type {
-    width: ${({ width, margin }) => (width + margin + LINE_WIDTH) / 2}px;
-  }
-
-  ${ItemContainer}:first-of-type &:first-of-type {
-    left: -${({ margin }) => (margin + LINE_WIDTH) / 2}px;
-    justify-content: flex-start;
-  }
-
-  ${ItemContainer}:last-of-type &:last-of-type {
-    justify-content: flex-end;
-  }
-`;
-
-const ReorderablePage = styled(ReorderableItem).attrs({ role: 'option' })`
-  display: flex;
-  z-index: 1;
-  margin-right: ${({ margin }) => margin}px;
-  ${ItemContainer}:last-of-type & {
-    margin-right: 0;
-  }
-`;
-
-const GridViewContainer = styled.section.attrs({
-  'aria-label': __('Grid View', 'web-stories'),
-})`
-  flex: 1;
-  margin: 70px 170px 70px 170px;
-  pointer-events: all;
-`;
-
-function Carousel() {
-  const {
-    pages,
-    currentPageId,
-    setCurrentPage,
-    arrangePage,
-  } = useStory(
-    ({
-      state: { pages, currentPageId },
-      actions: { setCurrentPage, arrangePage },
-    }) => ({ pages, currentPageId, setCurrentPage, arrangePage })
-  );
-  const { isRTL, version } = useConfig();
-  const [hasHorizontalOverflow, setHasHorizontalOverflow] = useState(false);
-  const [scrollPercentage, setScrollPercentage] = useState(0);
-  const [isGridViewOpen, setIsGridViewOpen] = useState(false);
-  const [workspaceWidth, setWorkspaceWidth] = useState(0);
-  const [carouselSize, setCarouselSize] = useState(0);
-
-  const listRef = useRef(null);
-  const pageRefs = useRef([]);
-  const wrapperRef = useRef(null);
-
-  const { isWideWorkspace, isVeryWideWorkspace } = useMemo(
-    () => ({
-      isWideWorkspace: workspaceWidth >= 700,
-      isVeryWideWorkspace: workspaceWidth >= 1000,
-    }),
-    [workspaceWidth]
-  );
-  const [pageThumbWidth, pageThumbHeight, pageThumbMargin] = useMemo(
-    () => (isWideWorkspace ? [40, 72, 16] : [36, 64, 16]),
-    [isWideWorkspace]
-  );
-  const menuWidth = useMemo(
-    () => (isVeryWideWorkspace ? MENU_WIDTH_VERYWIDE : MENU_WIDTH),
-    [isVeryWideWorkspace]
-  );
-  const carouselMaxWidth = useMemo(
-    // The carousel width has to match a whole number of page widths plus margins
-    () => {
-      const availableSpace = workspaceWidth - 2 * BUTTON_WIDTH - 2 * menuWidth;
-      const pageAndMargin = pageThumbWidth + pageThumbMargin;
-      const numPages = Math.floor(
-        (availableSpace - pageThumbMargin) / pageAndMargin
-      );
-      return pageThumbMargin + numPages * pageAndMargin;
-    },
-    [workspaceWidth, pageThumbWidth, pageThumbMargin, menuWidth]
-  );
-
-  const [resizedForPages, setResizedForPages] = useState(0);
-
-  const openModal = useCallback(() => setIsGridViewOpen(true), []);
-  const closeModal = useCallback(() => setIsGridViewOpen(false), []);
-
-  useResizeEffect(wrapperRef, (size) => setWorkspaceWidth(size.width), []);
-
-  useResizeEffect(
-    listRef,
-    (currentCarouselSize) => {
-      const { offsetWidth, scrollWidth, scrollLeft } = listRef.current;
-      const max = scrollWidth - offsetWidth;
-      setHasHorizontalOverflow(Math.ceil(scrollWidth) > Math.ceil(offsetWidth));
-      setScrollPercentage(scrollLeft / max);
-      setCarouselSize(currentCarouselSize);
-      setResizedForPages(pages.length);
-    },
-    [pages.length]
-  );
-
-  useLayoutEffect(() => {
-    if (hasHorizontalOverflow) {
-      const currentPageRef = pageRefs.current[currentPageId];
-
-      if (!currentPageRef || !currentPageRef.scrollIntoView) {
-        return;
-      }
-
-      currentPageRef.scrollIntoView({
-        inline: 'center',
-        behavior: 'smooth',
-      });
-    }
-  }, [currentPageId, hasHorizontalOverflow, carouselMaxWidth, pageRefs]);
-
-  useLayoutEffect(() => {
-    const rect = listRef.current.getBoundingClientRect();
-    setCarouselSize({ width: rect.width, height: rect.height });
-  }, []);
-
-  useLayoutEffect(() => {
-    const listElement = listRef.current;
-
-    const handleScroll = () => {
-      const { offsetWidth, scrollLeft, scrollWidth } = listElement;
-      const max = scrollWidth - offsetWidth;
-      setScrollPercentage(scrollLeft / max);
-    };
-
-    listElement.addEventListener('scroll', handleScroll, { passive: true });
-
-    return () => {
-      listElement.removeEventListener('scroll', handleScroll);
-    };
-  }, [hasHorizontalOverflow]);
-
-  const handleClickPage = (page) => () => setCurrentPage({ pageId: page.id });
-
-  const scrollBy = useCallback(
-    (offset) => {
-      if (isRTL) {
-        offset *= -1;
-      }
-
-      if (!listRef.current.scrollBy) {
-        listRef.current.scrollLeft += offset;
-        return;
-      }
-
-      listRef.current.scrollBy({
-        left: offset,
-        behavior: 'smooth',
-      });
-    },
-    [listRef, isRTL]
-  );
-
-  const scrollByPx = carouselSize.width;
-  const isAtBeginningOfList = isRTL
-    ? 1 === scrollPercentage
-    : 0 === scrollPercentage;
-  const isAtEndOfList = isRTL ? 0 === scrollPercentage : 1 === scrollPercentage;
-
-  const PrevButton = isRTL ? RightArrow : LeftArrow;
-  const NextButton = isRTL ? LeftArrow : RightArrow;
-
-  const rearrangePages = useCallback(
-    (oldPos, newPos) => {
-      const pageId = pages[oldPos].id;
-      arrangePage({ pageId, position: newPos });
-      setCurrentPage({ pageId });
-    },
-    [pages, arrangePage, setCurrentPage]
-  );
-
-  useCarouselKeys(wrapperRef, pageRefs, isRTL);
-
-  return (
-    <>
-      <Wrapper
-        menuWidth={menuWidth}
-        ref={wrapperRef}
-        data-testid="PageCarousel"
-        aria-label={__('Page Carousel', 'web-stories')}
-        data-ready={resizedForPages === pages.length}
-      >
-        <NavArea
-          area="prev-navigation"
-          marginBottom={CAROUSEL_BOTTOM_SCROLL_MARGIN}
-        >
-          <PrevButton
-            disabled={!hasHorizontalOverflow || isAtBeginningOfList}
-            onClick={() => scrollBy(-scrollByPx)}
-            width="24"
-            height="24"
-            aria-label={__('Scroll Backward', 'web-stories')}
-          />
-        </NavArea>
-        <PageList
-          maxWidth={carouselMaxWidth}
-          ref={listRef}
-          hasHorizontalOverflow={hasHorizontalOverflow}
-          aria-label={__('Pages List', 'web-stories')}
-          onPositionChange={rearrangePages}
-          mode={'horizontal'}
-          getItemSize={() => pageThumbWidth}
-        >
-          {pages.map((page, index) => {
-            const isCurrentPage = page.id === currentPageId;
-            const isInteractive = pages.length > 1;
-
-            return (
-              <ItemContainer
-                key={page.id}
-                ref={(el) => {
-                  pageRefs.current[page.id] = el;
-                }}
-              >
-                {index === 0 && (
-                  <PageSeparator
-                    position={0}
-                    width={pageThumbWidth}
-                    height={pageThumbHeight}
-                    margin={pageThumbMargin}
-                  >
-                    <Line height={pageThumbHeight} />
-                  </PageSeparator>
-                )}
-                <ReorderablePage position={index} margin={pageThumbMargin}>
-                  <PagePreview
-                    tabIndex={isCurrentPage && isInteractive ? 0 : -1}
-                    onClick={handleClickPage(page)}
-                    role="option"
-                    data-page-id={page.id}
-                    aria-label={
-                      isCurrentPage
-                        ? sprintf(
-                            /* translators: %s: page number. */
-                            __('Page %s (current page)', 'web-stories'),
-                            index + 1
-                          )
-                        : sprintf(
-                            /* translators: %s: page number. */
-                            __('Go to page %s', 'web-stories'),
-                            index + 1
-                          )
-                    }
-                    isActive={isCurrentPage && isInteractive}
-                    index={index}
-                    width={pageThumbWidth}
-                    height={pageThumbHeight}
-                    isInteractive={isInteractive}
-                  />
-                </ReorderablePage>
-                <PageSeparator
-                  position={index + 1}
-                  width={pageThumbWidth}
-                  height={pageThumbHeight}
-                  margin={pageThumbMargin}
-                >
-                  <Line height={pageThumbHeight} />
-                </PageSeparator>
-              </ItemContainer>
-            );
-          })}
-        </PageList>
-        <NavArea
-          area="next-navigation"
-          marginBottom={CAROUSEL_BOTTOM_SCROLL_MARGIN}
-        >
-          <NextButton
-            disabled={!hasHorizontalOverflow || isAtEndOfList}
-            onClick={() => scrollBy(scrollByPx)}
-            width="24"
-            height="24"
-            aria-label={__('Scroll Forward', 'web-stories')}
-          />
-        </NavArea>
-        <MenuArea>
-          <MenuIconsWrapper>
-            <OverflowButtons>
-              <KeyboardShortcutsMenu />
-            </OverflowButtons>
-            <WithTooltip
-              title={__('Grid View', 'web-stories')}
-              placement="left"
-            >
-              <StyledGridViewButton
-                onClick={openModal}
-                aria-label={__('Grid View', 'web-stories')}
-              />
-            </WithTooltip>
-          </MenuIconsWrapper>
-        </MenuArea>
-      </Wrapper>
-      <EditorVersion>
-        {sprintf(
-          /* translators: %s: editor version. */
-          __('Version %s', 'web-stories'),
-          version
-        )}
-      </EditorVersion>
-      <Modal
-        open={isGridViewOpen}
-        onClose={closeModal}
-        contentLabel={__('Grid View', 'web-stories')}
-        overlayStyles={{
-          alignItems: 'flex-start',
-        }}
-        contentStyles={{
-          pointerEvents: 'none',
-          flex: 1,
-        }}
-      >
-        <GridViewContainer>
-          <PlainStyled onClick={closeModal}>
-            {__('Back', 'web-stories')}
-          </PlainStyled>
-          <GridView />
-        </GridViewContainer>
-      </Modal>
-    </>
-  );
-}
-
-export default Carousel;
