+++ assets/src/edit-story/components/carousel/carouselPage.js
@@ -1,0 +1,171 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import styled from 'styled-components';
+import PropTypes from 'prop-types';
+
+/**
+ * WordPress dependencies
+ */
+import { __, sprintf } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+import { ReorderableSeparator, ReorderableItem } from '../reorderable';
+import PagePreview, { THUMB_FRAME_HEIGHT } from './pagepreview';
+import useCarousel from './useCarousel';
+
+const LINE_WIDTH = 4;
+
+const Line = styled.div`
+  background: ${({ theme }) => theme.colors.accent.primary};
+  height: ${({ height }) => height - THUMB_FRAME_HEIGHT}px;
+  width: ${LINE_WIDTH}px;
+  margin: 0px;
+`;
+
+const ItemContainer = styled.div`
+  display: flex;
+  position: relative;
+`;
+
+const PageSeparator = styled(ReorderableSeparator)`
+  position: absolute;
+  bottom: 0;
+  left: ${({ width }) => width / 2}px;
+  width: ${({ width, margin }) => width + margin}px;
+  height: ${({ height }) => height - THUMB_FRAME_HEIGHT}px;
+  display: flex;
+  justify-content: center;
+
+  ${ItemContainer}:first-of-type &:first-of-type,
+  ${ItemContainer}:last-of-type &:last-of-type {
+    width: ${({ width, margin }) => (width + margin + LINE_WIDTH) / 2}px;
+  }
+
+  ${ItemContainer}:first-of-type &:first-of-type {
+    left: -${({ margin }) => (margin + LINE_WIDTH) / 2}px;
+    justify-content: flex-start;
+  }
+
+  ${ItemContainer}:last-of-type &:last-of-type {
+    justify-content: flex-end;
+  }
+`;
+
+const ReorderablePage = styled(ReorderableItem).attrs({ role: 'option' })`
+  display: flex;
+  z-index: 1;
+  margin-right: ${({ margin }) => margin}px;
+`;
+
+function CarouselPage({ pageId, index }) {
+  const {
+    pageThumbWidth,
+    pageThumbHeight,
+    pageThumbMargin,
+    clickPage,
+    setPageRef,
+    page,
+    isCurrentPage,
+    hasMultiplePages,
+  } = useCarousel(
+    ({
+      state: {
+        pageThumbWidth,
+        pageThumbHeight,
+        pageThumbMargin,
+        numPages,
+        pages,
+        currentPageId,
+      },
+      actions: { clickPage, setPageRef },
+    }) => ({
+      pageThumbWidth,
+      pageThumbHeight,
+      pageThumbMargin,
+      clickPage,
+      setPageRef,
+      page: pages.find(({ id }) => id === pageId),
+      isCurrentPage: pageId === currentPageId,
+      hasMultiplePages: numPages > 1,
+    })
+  );
+
+  if (!page || !page.id) {
+    return null;
+  }
+
+  return (
+    <ItemContainer key={page.id} ref={(el) => setPageRef(page, el)}>
+      {index === 0 && (
+        <PageSeparator
+          position={0}
+          width={pageThumbWidth}
+          height={pageThumbHeight}
+          margin={pageThumbMargin}
+        >
+          <Line height={pageThumbHeight} />
+        </PageSeparator>
+      )}
+      <ReorderablePage position={index} margin={pageThumbMargin}>
+        <PagePreview
+          tabIndex={isCurrentPage && hasMultiplePages ? 0 : -1}
+          onClick={() => clickPage(page)}
+          role="option"
+          data-page-id={page.id}
+          aria-label={
+            isCurrentPage
+              ? sprintf(
+                  /* translators: %s: page number. */
+                  __('Page %s (current page)', 'web-stories'),
+                  index + 1
+                )
+              : sprintf(
+                  /* translators: %s: page number. */
+                  __('Go to page %s', 'web-stories'),
+                  index + 1
+                )
+          }
+          isActive={isCurrentPage && hasMultiplePages}
+          page={page}
+          width={pageThumbWidth}
+          height={pageThumbHeight}
+          isInteractive={hasMultiplePages}
+        />
+      </ReorderablePage>
+      <PageSeparator
+        position={index + 1}
+        width={pageThumbWidth}
+        height={pageThumbHeight}
+        margin={pageThumbMargin}
+      >
+        <Line height={pageThumbHeight} />
+      </PageSeparator>
+    </ItemContainer>
+  );
+}
+
+CarouselPage.propTypes = {
+  pageId: PropTypes.string.isRequired,
+  index: PropTypes.number.isRequired,
+};
+
+export default CarouselPage;
